Add distributed-read of an ExodusII file + partitioning with Zoltan2
This can be tested by passing an exodus mesh filename as the first
command line argument. A simple mesh is in tests/box_24.exo. Paraview
should be able to visualize that mesh by 'paraview box_24.exo'. Example:

./charmrun +p2 Main/exam2m ../../tmp/box_24.exo

Note that an intel build does not play nice with system-wide
HDF5/NetCDF, it yields a runtime segfault, similar to:

===================================================================================
=   BAD TERMINATION OF ONE OF YOUR APPLICATION PROCESSES
=   RANK 0 PID 22228 RUNNING AT euler
=   KILLED BY SIGNAL: 11 (Segmentation fault)
===================================================================================

This can be remedied by ignoring the system-wide HDF5/NetCDF by passing
to the TPL-build cmake:

-DCMAKE_DISABLE_FIND_PACKAGE_HDF5=true -DCMAKE_DISABLE_FIND_PACKAGE_NetCDF=true

Next is to generate some field data on the mesh at nodes and write it
out to a file.