Transfer interpolated solution values back to dest mesh
Previously we had switched the source and dest mesh in doing collision
detection. Here we change it so the vertices of dest are collided against the
tets of the source mesh. Beyond that, once actual collisions are determined,
the source mesh determines the solution value at the dest vertices and sends
that information back to the destination mesh.