Add MKL/LAPACK for Exodus/NetCDF
MKL or LAPACK is required for Exodus/NetCDF/HDF5 and by default if found
on the system, it was not always found and linked when built as a
library. This ensures that MKL, and if not found, LAPACK, is searched and
used if it is found, and if not found, it is built by TPL build from
external.