Virtualization (#19)
* Ensure that duplicate points are not sent for collision detection

For vertices on the boundaries, multiple chares have these vertices
in their mesh data. This change designates the chare with the lowest
index as the owner of these points and ensures that it is the only
chare sending them to collision detection.

* Write new output instead of writing all meshes into a single file

* Initialize non-collision dest nodes and fix interpolation of source tet

Comment out ownership detection for now.

* Add 0.2 x 0.9 sphere mesh

* Re-enable checking ownership

Only has an effect in parallel. This will likely not be correct, since
now the dest_index will not map to the correct one.

* Port over Quinoa cmake code for testing output with exodiff

* Add sphere2box regression test

Also added a ParaView state file for helping quick visualization of the
results.

* Correct when background solution values are set

Also add regression test with sphere to box on 2PEs.

Also add ParaView state file for quickly visualizing the 2PE sphere2box
solution.

* Add virtualization

The third, optional, cmd line argument can now specify a real number
between 0.0 and 1.0, both inclusive to configure virtualization.

Also add a regression test that exercises this.

* Force our HDF5 instead of ubuntu's

I use the system-wide on debian/stable, but on ubuntu, the system-wide
hdf5 sometimes yields this very useful error message from Zoltan:

No space on proc 0 - number of bytes requested = 18446744068053123104

* Try latest ubuntu for CI

* Use system-wide HDF5 for CI

* Store mapping between actual vertex index and collision vertex index

* Shorten comment fit 80 cols

* Update baseline due to ownership map change

* Enable ownership check in parallel

So that it passes the parallel regression test that now has an updated
baseline, incorporating ownership.

* Add condition for not finding nodeset in Worker::owner

This fixes the segfault seen previously in the overdecomposition test

* Remove ownership

Will return to it if/when necessary.

Co-authored-by: Eric Mikida <epmikida@hpccharm.com>
Co-authored-by: Nitin Bhat <nitin@hpccharm.com>
Co-authored-by: Nitin Bhat <nbhat4@illinois.edu>